+++ tests/backends/logging/test_graylog.py
@@ -1,7 +1,13 @@
 """Tests for Ralph graylog storage backend"""
 
 from ralph.backends.logging.graylog import GraylogLogging
-from ralph.defaults import RALPH_GRAYLOG_HOST, RALPH_GRAYLOG_PORT
+from ralph.defaults import (
+    RALPH_GRAYLOG_ADMIN_PASSWORD,
+    RALPH_GRAYLOG_ADMIN_USERNAME,
+    RALPH_GRAYLOG_EXTERNAL_PORT,
+    RALPH_GRAYLOG_HOST,
+    RALPH_GRAYLOG_PORT,
+)
 
 
 def test_backends_logging_graylog_logging_instantiation():
@@ -11,8 +17,12 @@ def test_backends_logging_graylog_logging_instantiation():
     storage = GraylogLogging(
         host=RALPH_GRAYLOG_HOST,
         port=RALPH_GRAYLOG_PORT,
+        external_port=RALPH_GRAYLOG_EXTERNAL_PORT,
+        username=RALPH_GRAYLOG_ADMIN_USERNAME,
+        password=RALPH_GRAYLOG_ADMIN_PASSWORD,
     )
 
     assert storage.name == "graylog"
     assert storage.host == "graylog"
     assert storage.port == 12201
+    assert storage.external_port == 9000